% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_max.R
\name{p_rank_max}
\alias{p_rank_max}
\title{Percentile Max:  This function formulas is used if the best rate for a given indicator is 'max'. For example, high school graduation.}
\usage{
p_rank_max(a)
}
\arguments{
\item{Calculates}{a percentile value}
}
\value{
A Percentile value
}
\description{
Percentile Max:  This function formulas is used if the best rate for a given indicator is 'max'. For example, high school graduation.
}
\examples{
data <- data.frame (county_id= c("06037", "06001", "06113", "06101", "06091"),
                    county_name = c("Los Angeles", "Alameda", "Yolo", "Sutter", "Sierra"),
                    total_rate = c("35.70", "40.03", "48.99", "82.93", "92.55"))

data$performance_pctile <-  p_rank_max(data$total_rate)

}
