% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_min.R
\name{p_rank_min}
\alias{p_rank_min}
\title{Percentile Min: This function formula is used if the best rate for a given indicator is 'min.' For example, chronic absenteeism.}
\usage{
p_rank_min(a)
}
\arguments{
\item{Calculates}{a percentile value}
}
\value{
A Percentile value
}
\description{
Percentile Min: This function formula is used if the best rate for a given indicator is 'min.' For example, chronic absenteeism.
}
\examples{
Calculates the percentile rank max if the 'best' is 'max'. Otherwise, it calculates the percentile of the minimum.
data <- data.frame (county_id= c("06037", "06001", "06113", "06101", "06091"),
                    county_name = c("Los Angeles", "Alameda", "Yolo", "Sutter", "Sierra"),
                    total_rate = c("35.70", "40.03", "48.99", "82.93", "92.55"))

data$performance_pctile <-  p_rank_min(data$total_rate)
}
