% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile-package.R
\docType{package}
\name{percentile-package}
\alias{percentile}
\alias{percentile-package}
\title{percentile: Calculates percentile depending if the best is 'max' or 'min'}
\description{
This is a handy package that stores the percentile functions depending if the best is 'max' or 'min.'
}
\author{
\strong{Maintainer}: David Segovia \email{dsegovia@catalystcalifornia.org} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
